feat(UX active buffers): close display and search prompt on empty pattern
When the search prompt is empty and we exit it with <ESC> instead of
closing the search prompt and the active buffer display, we merely
close the former while entering the latter. This is the same behavior as
invoking <CR> on an empty search pattern. We change the UX so that if
the search prompt is empty, <ESC> now closes both, the search prompt
*and* the active buffer display.