fix: bug causing the cmdline mirroring to error
We attempt to fix a bug where the mirroring of the cmdline would error.

The following command reproduces the error reliably:

1. search using vimgrep for anything that can be found in your project

2. repeat

On the second `/` of $ vimgrep /pattern/ in the second step the plugin
errors complaining about an `Invalid lnum`. The bug is likely caused due
to an expansion of the search history silently invoked by vimgrep. This
search history may consist of a long buffer leading to the `Invalid
lnum` error.

After some iterations we have not yet found a clean fix for this.

What we have tried so far:
1. clean state.buf and state.mirror
2. defer/debounce the input
3. use `vim.api.nvim_buf_set_text` instead of
   `vim.api.nvim_buf_set_lines`

Eventually, we call `update_cmdline` inside a pcall and skip the update
on error. You can see this by using :messages (at debug level).